debug(godmode): Basic implementation of godmode debugged.
Restore human form now no longer crashes the save game load.
Moved it to the very end of the file.
.
Gave up on increase of Mana, Stamina and Health.
Fail to understand the format, even when taking a close
look at the bit list of save games where these values are known.