fix(godmode): Got god mode to work. Including disabling it.
Disabling now preserves stat and skill points won during god
mode operation.
.
Sadly I was compelled to create an external .humanity file for
restoring action after all. TODO: Add option to the command line
for actively selecting it. As it stands it is needed to be
beside the save game file (where it is created).
.
Getting late. Creation of UI and Wiki page pending.