feat(hf): Now proper support of Quarter-pseudo-floats in HMS.
Wrote some HMS_parsing and encoding functions in Data.
Finally understood the format of HP Mana and Stamina values.
.
Highest 8 bits encode in their most significant pair the number
of quarter points.
Lowest 13 bits encode the actual integer value.
Btw.: This answers the question of the highest HP possible in D2:
It is 2^13 == 8192.