feat(cube create): Added --ensure_horadric.
Will verify if there is a cube in the total inventory.
If so: No effect. Else, a new cube will be created and
added to the top-left corner of the character inventory.
Any objects that are in the way will be shifted inside the
cube. Geometries of stashes are such that this will work
in any case. The cube's interior is large enough.