feat(Data): Allowing int values for progression setter.
Had a theory, that progression indicates if a movie will be
shown when going east. Progression in Nightmare Lut Gholein is 6
for instance.
.
A first test (setting progress back to 5, starting out in Lut Gholein,
going back and then using Warriv to go east again, could not prove
this though.
.
Anyway, progression is now part of the info text, too.