feat(jewelize): Found that inferior items may have runewords.
And what is more: Their inferiority is not encoded in extended mods
but only in the quality attributes section as a 3 bit integer,
from which their lower stats are probably computed in a fixed,
simple formula.
.
Be that as it may: Having no extended mods makes runeworded
inferior items ideal candidates for transmogrification using the
jewelize function. All that was necessary, was making such items
eligible for the process. Did just that.