fix(corpse): Fixed read-in of corpse items.
Sadly, the Necromancer, Charonson has died. On the upside, this
raises opportunity to test the code using a dead character.
Which failed. The header is really 20 bytes long (as opposed
to 16), since the first item beyond it is 'JM..' where .. is a
2 byte item counter. Fixed a bug in Item.get_block_index() to this end.
.
Also, added dead Charonson as an example .d2s file.