feat(waypoints): Completed waypoint setter.
Including fixes in the quests datastructures to enable
opening the tabs with the activated waypoints, even, if
necessary quests have not yet been concluded.
.
There is one notable bug: While this technique seems to
work for all five acts, if act V is unlocked for waypoint
access, the tab selector for act V will remeain invisible
within the waypoint menu. However, it is clickable all the same!
.
Also added 'enable nirvana' for creating a level 86 millionaire
character who holds the rank of Guradian, Matriarch or Patriarch.
.
Next Steps
==========
So far, waypoint setting and nirvana mode are availbale only
from the command line. The GUI will need extension for the new features.
.
Another idea is allowing to create small charms for
teleport and breath of the dying ('cause its cool!) within
the rune forge in Horazon's folly.