feat(hf): Added Checkbox 'toggle waypoint Halls of Pain'.
The practical usecase of being able to toggle this waypoint
(only active, when Anya has given the 'Betrayal in Harrogath' quest)
is, that it, in reciprocal effect, also toggles the orange
portal opened by Anya. Turning this waypoint off will allow to
easily enter the entrance of Nihlathak's temple again.
This is interesting for Necromancers (free bodies!) and people
who like Pindlespin.