feat(hf): Added "Enable Nirvana" button.
Promotes to a credible level 86 character. As an additional
touch I would also like to activate all waypoints of normal,
nightmare and hell (and similar functionality for enable nightmare
and enable hell that get all waypoints for normal, resp. normal
and nightmare).
.
However, this still has issues with quest logic and suffers from the
"tab 5" bug. So for now I will not activate that beyond the console
interface options.